Remove unused interrupt enable flag
According to the datasheet, bit 6 of FIFO_IENABLE_REG "enables all 6
status conditions to interrupt" (not to be confused with
FIFO_IENABLE_ALL which is just a bitfield with all individual conditions
set). I could not reproduce this on actual hardware. I'm assuming it is
a documentation error, and removing this special case from the handler.